requirements.txt
tqdm
Pillow
argparse
https://cdn.evilmadscientist.com/dl/ad/public/AxiDraw_API.zip